chatgpt_package/requirements.txt
# Gap Scanner Project Dependencies
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
tqdm>=4.64.0

# Environment configuration (REQUIRED for security)
python-dotenv>=1.0.0

# Financial data APIs
alpaca-trade-api>=3.0.0

# HTTP and networking
urllib3>=1.26.0
certifi>=2022.0.0

# Date and time handling
python-dateutil>=2.8.0

# Data serialization
toml>=0.10.2

# Optional: Enhanced functionality
# yfinance>=0.2.0  # Yahoo Finance data (if needed)
# beautifulsoup4>=4.11.0  # Web scraping (if needed)

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=5.0.0
# black>=22.0.0
# isort>=5.10.0

# Documentation
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0

# Note:
# - sqlite3 is included with Python standard library
# - ThetaData requires separate terminal installation
# - Some APIs may require additional authentication setup
streamlit>=1.36.0

